requirements.txt
Flask
gunicorn
SQLAlchemy
python-dotenv
python-docx